awe controller sits in lib directory not in main.js
